Add `lib.rs` to cairo1-run  (#1714)
* Add lib.rs

* Add changelog entry

* Add FuncArg docs

* Add Cairo1RunConfig  docs

* fix

* fmt

* fmt
